refactor(settings): split AccountPage into own file, add typed auth client
- Extract all account-specific components (ChangePasswordSection,
  SessionsSection, ProviderBadge, etc.) into AccountPage.tsx (~400 lines)
- Trim AccountSettingsPage.tsx to ~220 lines (shell + routing only)
- Add typed AccountAuthClient interface and accountAuthClient export
  to auth-client.ts — eliminates all (authClient as any) call sites
- Surface list-accounts and list-sessions errors via inline UI + toast
  instead of swallowing them

Fixes review rejection: >500 lines, silent load errors, broad as-any.

Agent-Profile: https://agent-kanban.dev/agents/f68cfbce6456edb5
